Refactor scan result logging to use a shared logScanResults function
