+++ src/realtimeScanners/scanners/iac/iacScannerService.ts
@@ -13,6 +13,7 @@ import fs from "fs";
 import { minimatch } from "minimatch";
 import { CxRealtimeEngineStatus } from "@checkmarxdev/ast-cli-javascript-wrapper/dist/main/oss/CxRealtimeEngineStatus";
 import { IgnoreFileManager } from "../../common/ignoreFileManager";
+import {logScanResults} from "../common";
 
 export class IacScannerService extends BaseScannerService {
 	private diagnosticsMap = new Map<string, vscode.Diagnostic[]>();
@@ -130,7 +131,7 @@ export class IacScannerService extends BaseScannerService {
 				fullScanResults = await cx.iacScanResults(tempFilePath, this.getContainersManagementTool(), "");
 			}
 
-			this.logScanResults("iac", fullScanResults);
+			logScanResults("iac", fullScanResults);
 
 			ignoreManager.removeMissingIac(fullScanResults, filePath);
